requirements.txt
Django>=5.1,<5.3
docxtpl>=0.17
python-pptx>=1.0.2
Pillow>=10.4
psycopg[binary]>=3.2
celery>=5.4
redis>=5.2
